move localStorage module
